fix: use temporary dir for package manager files
Instead of relying on a fixed directory in /tmp that may or may not
exist to write package manager logs to, use the tempfile library to
create a temporary directory within the container for these at runtime.

Also, move the functions in installer.py to a new Installer class in
order to store and reuse the buildah container and mount names, the
package manager, as well as the temporary directory which is created on
init.