fix: create tempdir outside of container image-build creates
If an error occurs in image-build, the buildah container it creates is
unmounted and deleted. This can be rather unhelpful if one wants to
examine package manager logs to find out more about any errors that may
have occurred. Thus, this commit causes the installer to create the
image-build tempdir (where the package manager cache/log files get
stored) on the machine (host or container) where image-build is running.
That way even if/when the buildah container is destroyed, the tempdir
with the logs, etc. remains for inspection.