fix(publish): be smarter and verbose about getting initramfs path
Instead of blindly choosing the first kernel version in /lib/modules and
using that to set the initramfs path, use a loop to see if the initramfs
actually exists and iterate through the versions until one is found. If
none is found, raise an exception. This fixes the 'initrd' referenced
before assignment errors that occur without context if the initramfs is
not found.

This should add more info for use in troubleshooting initramfs
generation.

Signed-off-by: Devon Bautista <17506592+synackd@users.noreply.github.com>